changed authentication for order mutations
addOrder - owner, updateOrder - owner/walker, updateOrderStatus - owner/walker, removeOrder - admin

+++ server/schemas/resolvers.js
@@ -283,35 +283,52 @@ const resolvers = {
            - removeOrder
         */
         addOrder: async (parent, { input }, context) => {
-            const order = await Order.create(input);
+            // only owner can add order
+            if (context.owner) {
+                const order = await Order.create(input);
 
-            return order;
+                return order;
+            }
+      
+            throw new AuthenticationError('Not logged in');
         },
 
         updateOrder: async (parent, {order_id, input }, context) => {
-            const order = await Order.findByIdAndUpdate(
-                order_id, 
-                input,
-                { new: true, runValidators: true }
-            );
+            if (context.owner || context.walker) {
+                const order = await Order.findByIdAndUpdate(
+                    order_id, 
+                    input,
+                    { new: true, runValidators: true }
+                );
 
-            return order;
+                return order;
+            }
+      
+            throw new AuthenticationError('Not logged in');
         },
 
         updateOrderStatus: async (parent, {order_id, status }, context) => {
-            const order = await Order.findByIdAndUpdate(
-                order_id, 
-                {status: status},
-                { new: true, runValidators: true }
-            );
-        
-            return order;
+            if (context.owner || context.walker) {
+                const order = await Order.findByIdAndUpdate(
+                    order_id, 
+                    {status: status},
+                    { new: true, runValidators: true }
+                );
+            
+                return order;
+            }
+      
+            throw new AuthenticationError('Not logged in');
         },
 
         removeOrder: async (parent, {order_id, input }, context) => {
-            const order = await Order.findByIdAndDelete(order_id);
+            if (context.owner && context.owner.admin) {
+                const order = await Order.findByIdAndDelete(order_id);
 
-            return order;
+                return order;
+            }
+        
+            throw new AuthenticationError('Not logged in');
         },
 
         /* Review mutations
